backend/requirement.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
cohere>=4.9.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
fastapi
uvicorn
cohere
qdrant-client
openai
python-dotenv
